moved results business logic to parent Results component
